@@ -7,10 +7,9 @@
     "arr-flatten": "^1.1.0",
     "arr-union": "^3.1.0",
     "array-unique": "^0.3.2",
-    "react": "^18.3.1",
-    "react-addons-update": "^15.6.3",
-    "react-beautiful-dnd": "^13.1.1",
-    "react-dom": "^18.3.1",
+    "react": "^19.0.0",
+    "@hello-pangea/dnd": "^16.5.0",
+    "react-dom": "^19.0.0",
     "web-vitals": "^4.2.4"
   },
   "devDependencies": {
